Pre-migracion de funciones
